requirements.txt
# LDMR算法仿真系统依赖
# 核心科学计算库
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 图计算和网络分析
networkx>=2.8.0

# 可视化库
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 配置管理
PyYAML>=6.0

# 数据处理
scikit-learn>=1.0.0

# 可选：性能优化
numba>=0.56.0
